Added RuboCop with Foreman's Configuration
RuboCop is a Ruby static code analyzer and code formatter.
It will enforce many of the guidelines outlined in the community:
https://rubystyle.guide/

About RuboCop:
https://docs.rubocop.org/rubocop/1.37/index.html
How to install:
https://sourcelevel.io/blog/rubocop-how-to-install-and-configure Added RuboCop

RuboCop is a Ruby static code analyzer and code formatter.
It will enforce many of the guidelines outlined in the community:
https://rubystyle.guide/

About RuboCop:
https://docs.rubocop.org/rubocop/1.37/index.html
How to install:
https://sourcelevel.io/blog/rubocop-how-to-install-and-configure Added RuboCop

RuboCop is a Ruby static code analyzer and code formatter.
It will enforce many of the guidelines outlined in the community:
https://rubystyle.guide/

About RuboCop:
https://docs.rubocop.org/rubocop/1.37/index.html
How to install:
https://sourcelevel.io/blog/rubocop-how-to-install-and-configure